Scaffold was broken so removing the login ugly thing with js now (forgot to push so late push)

+++ FiveAnime/Program.cs
@@ -18,11 +18,9 @@ namespace FiveAnime
 
         public static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                    webBuilder.UseStartup<Startup>()
-                .UseKestrel(options =>
+                .ConfigureWebHostDefaults(webBuilder => webBuilder.UseStartup<Startup>().UseKestrel(options =>
                 {
-                    options.Limits.MaxRequestBodySize = int.MaxValue;
+                    options.Limits.MaxRequestBodySize = long.MaxValue;
                 }));
     }
 }
